pyproject.toml
[tool.poetry]
name = "citas-dinissan"
version = "0.7.0"
description = ""
authors = ["Eglenelid Gamaliel Gutierrez Hernandez <eglenelid.gamaliel@gmail.com>"]
readme = "README.md"
packages = [{include = "citas_dinissan"}]

[tool.poetry.dependencies]
python = "^3.8.0"
django = "^4.0.0"
django-q = "^1.3.9"
djangorestframework = "^3.14.0"
django-drf-filepond = "^0.5.0"

aiohttp = "^3.8.3"
httpx = "^0.23.1"
requests = "^2.28.1"
numpy = "^1.24.1"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
mssql-django = "^1.2"
black = "^22.12.0"
pylint = "^2.15.9"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
